+++ src/Network/Network.Test/ScenarioTests/AzureFirewallPolicyTests.ps1
@@ -1610,7 +1610,7 @@ function Test-AzureFirewallPolicyExplicitProxyCRUD {
         $explicitProxySettings = New-AzFirewallPolicyExplicitProxy -EnableExplicitProxy  -HttpPort 85 -EnablePacFile  -PacFilePort 122 -PacFile $pacFile
 
         # Create AzureFirewallPolicy (with Explicit Proxy Settings)
-        $azureFirewallPolicy = New-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -ExplicitProxy $explicitProxySettings -UserAssignedIdentityIds @($userAssignedManagedIdentity)
+        $azureFirewallPolicy = New-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -ExplicitProxy $explicitProxySettings -UserAssignedIdentityId @($userAssignedManagedIdentity)
 
         # Get AzureFirewallPolicy
         $getAzureFirewallPolicy = Get-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname
@@ -1697,7 +1697,7 @@ function Test-AzureFirewallPolicyWithMultipleUAMI {
         # Two different UAMIs for Explicit Proxy and TLS Inspection
         $exProxy = New-AzFirewallPolicyExplicitProxy -EnableExplicitProxy  -HttpPort 86 -EnablePacFile  -PacFilePort 124 -PacFile $pacFile
         # Set AzureFirewallPolicy
-        $azureFirewallPolicy = Set-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -SkuTier $tier -ExplicitProxy $exProxy -TransportSecurityName $transportSecurityName -TransportSecurityKeyVaultSecretId $transportSecurityId -UserAssignedIdentityIds @($userAssignedManagedIdentity, $transportSecurityIdentity)
+        $azureFirewallPolicy = Set-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -SkuTier $tier -ExplicitProxy $exProxy -TransportSecurityName $transportSecurityName -TransportSecurityKeyVaultSecretId $transportSecurityId -UserAssignedIdentityId @($userAssignedManagedIdentity, $transportSecurityIdentity)
 
         $getAzureFirewallPolicy = Get-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname
 
@@ -1713,7 +1713,7 @@ function Test-AzureFirewallPolicyWithMultipleUAMI {
         # Single UAMI to set both Explicit Proxy and TLS Inspection
         $exProxy = New-AzFirewallPolicyExplicitProxy -EnableExplicitProxy  -HttpPort 86 -EnablePacFile  -PacFilePort 124 -PacFile $pacFile
         # Set AzureFirewallPolicy
-        $azureFirewallPolicy = Set-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -SkuTier $tier -ExplicitProxy $exProxy -TransportSecurityName $transportSecurityName -TransportSecurityKeyVaultSecretId $transportSecurityId -UserAssignedIdentityIds @($userAssignedManagedIdentity)
+        $azureFirewallPolicy = Set-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -SkuTier $tier -ExplicitProxy $exProxy -TransportSecurityName $transportSecurityName -TransportSecurityKeyVaultSecretId $transportSecurityId -UserAssignedIdentityId @($userAssignedManagedIdentity)
 
         $getAzureFirewallPolicy = Get-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname
 
@@ -1724,22 +1724,6 @@ function Test-AzureFirewallPolicyWithMultipleUAMI {
         Assert-AreEqual "UserAssigned" $getAzureFirewallPolicy.Identity.type
         Assert-AreEqual 1 $getAzureFirewallPolicy.Identity.userAssignedIdentities.Count
         Assert-True { $getAzureFirewallPolicy.Identity.userAssignedIdentities.ContainsKey($userAssignedManagedIdentity) }
-
-        #Pass Two UAMI for Explicit Proxy and TLS via both parameters
-        $exProxy = New-AzFirewallPolicyExplicitProxy -EnableExplicitProxy  -HttpPort 86 -EnablePacFile  -PacFilePort 124 -PacFile $pacFile
-        # Set AzureFirewallPolicy
-        $azureFirewallPolicy = Set-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname -Location $location -SkuTier $tier -ExplicitProxy $exProxy -TransportSecurityName $transportSecurityName -TransportSecurityKeyVaultSecretId $transportSecurityId -UserAssignedIdentityIds @($userAssignedManagedIdentity) -UserAssignedIdentityId $transportSecurityIdentity
-
-        $getAzureFirewallPolicy = Get-AzFirewallPolicy -Name $azureFirewallPolicyName -ResourceGroupName $rgname
-
-        Assert-NotNull $getAzureFirewallPolicy.ExplicitProxy
-        Assert-AreEqual 86 $getAzureFirewallPolicy.ExplicitProxy.HttpPort
-        Assert-AreEqual 124 $getAzureFirewallPolicy.ExplicitProxy.PacFilePort
-        Assert-AreEqual $pacFile $getAzureFirewallPolicy.ExplicitProxy.PacFile
-        Assert-AreEqual "UserAssigned" $getAzureFirewallPolicy.Identity.type
-        Assert-AreEqual 2 $getAzureFirewallPolicy.Identity.userAssignedIdentities.Count
-        Assert-True { $getAzureFirewallPolicy.Identity.userAssignedIdentities.ContainsKey($userAssignedManagedIdentity) }
-        Assert-True { $getAzureFirewallPolicy.Identity.userAssignedIdentities.ContainsKey($transportSecurityIdentity) }
     }
     finally {
         # Cleanup