+++ src/Network/Network/AzureFirewallPolicy/NewAzureFirewallPolicyCommand.cs
@@ -131,14 +131,7 @@ namespace Microsoft.Azure.Commands.Network
             HelpMessage = "ResourceId of the user assigned identity to be assigned to Firewall Policy.")]
         [ValidateNotNullOrEmpty]
         [Alias("UserAssignedIdentity")]
-        public string UserAssignedIdentityId { get; set; }
-
-        [Parameter(
-            Mandatory = false,
-            HelpMessage = "Multiple ResourceId of the user assigned identities to be assigned to Firewall Policy.")]
-        [ValidateNotNullOrEmpty]
-        [Alias("UserAssignedIdentities")]
-        public string[] UserAssignedIdentityIds { get; set; }
+        public string[] UserAssignedIdentityId { get; set; }
 
         [Parameter(
             Mandatory = false,
@@ -208,23 +201,15 @@ namespace Microsoft.Azure.Commands.Network
                 firewallPolicy.Snat = this.Snat;
             }
 
-            if (this.UserAssignedIdentityId != null || this.UserAssignedIdentityIds != null)
+            if (this.UserAssignedIdentityId != null)
             {
                 var userAssignedIdentities = new Dictionary<string, PSManagedServiceIdentityUserAssignedIdentitiesValue>();
 
-                if (this.UserAssignedIdentityId != null)
+                foreach (var identityId in this.UserAssignedIdentityId)
                 {
-                    userAssignedIdentities.Add(this.UserAssignedIdentityId, new PSManagedServiceIdentityUserAssignedIdentitiesValue());
-                }
-
-                if (this.UserAssignedIdentityIds != null)
-                {
-                    foreach (var identityId in this.UserAssignedIdentityIds)
+                    if (!userAssignedIdentities.ContainsKey(identityId))
                     {
-                        if (!userAssignedIdentities.ContainsKey(identityId))
-                        {
-                            userAssignedIdentities.Add(identityId, new PSManagedServiceIdentityUserAssignedIdentitiesValue());
-                        }
+                        userAssignedIdentities.Add(identityId, new PSManagedServiceIdentityUserAssignedIdentitiesValue());
                     }
                 }
 
@@ -246,7 +231,7 @@ namespace Microsoft.Azure.Commands.Network
                     throw new ArgumentException("TransportSecurityName must be provided with TransportSecurityKeyVaultSecretId");
                 }
 
-                if (this.Identity == null && this.UserAssignedIdentityId == null && this.UserAssignedIdentityIds == null)
+                if (this.Identity == null && this.UserAssignedIdentityId == null)
                 {
                     throw new ArgumentException("Identity must be provided with TransportSecurityKeyVaultSecretId");
                 }