+++ src/Network/Network/ChangeLog.md
@@ -22,15 +22,13 @@
 * Added support to associate a DDoS custom policy with a Load Balancer frontend IP configuration.
     - Added the `-DdosCustomPolicyId` parameter to `New-AzLoadBalancerFrontendIpConfig`, `Add-AzLoadBalancerFrontendIpConfig`, and `Set-AzLoadBalancerFrontendIpConfig`.
     - Added the `-RemoveDdosCustomPolicy` switch to `Set-AzLoadBalancerFrontendIpConfig`.
+* Changed `UserAssignedIdentityId` parameter type from string to list of strings in Azure Firewall Policy to input multiple UAMIs
 
 ## Version 7.26.0
 * Onboarded `Microsoft.KubernetesConfiguration/privateLinkScopes` to Private Link Common Cmdlets
 * Added EdgeZone (Extended Location) parameter support for Azure Firewall
     - Added `-EdgeZone` parameter to `New-AzFirewall` cmdlet
     - When `-EdgeZone` is specified, availability zones are not supported for Azure Firewall resources
-* Added `UserAssignedIdentityIds` parameter in Azure Firewall Policy to input multiple UAMIs
-    - Added `-UserAssignedIdentityIds` parameter to `New-AzFirewallPolicy` and `Set-AzFirewallPolicy` cmdlets
-    - Supports assigning multiple User Assigned Managed Identities (e.g., separate identities for Explicit Proxy and TLS Inspection)
     
 ## Version 7.25.1
 * Onboarded `Microsoft.DataReplication/replicationVaults` to Private Link Common Cmdlets